+++ src/components/TreeFarm.vue
@@ -6,6 +6,7 @@
   /* eslint no-param-reassign: off */
   /* esint no-shadow: off */
   import 'three/examples/js/controls/OrbitControls';
+  import Tree from 'objects/Tree';
   import makeSprite from 'utils/makeSprite';
 
   var scene,
@@ -13,13 +14,6 @@
     renderer,
     orbit,
     light;
-
-  const treeMaterial = new THREE.MeshPhongMaterial({
-    color: 0x2c9e4b,
-    shininess: 20,
-    side: THREE.FrontSide,
-    flatShading: true,
-  });
 
   function onWindowResize() {
     camera.aspect = window.innerWidth / window.innerHeight;
@@ -47,39 +41,6 @@
     sprite.needsUpdate = true;
 
     return sprite;
-  }
-
-  class Cone extends THREE.Mesh {
-    constructor(material, size = 10, translate) {
-      const geometry = new THREE.CylinderGeometry(size / 2, size, size, 6);
-      if (translate) {
-        geometry.applyMatrix(new THREE.Matrix4().makeTranslation(0, size, 0));
-      }
-      super(geometry, material);
-    }
-  }
-
-  class Tree extends THREE.Object3D {
-    constructor(size = 6 + Math.random()) {
-      super();
-      let lastCone;
-      let cone;
-
-      for (let i = 0; i < size; i++) {
-        cone = new Cone(treeMaterial, (size - i) + 1, i);
-        cone.position.y = 0;
-        if (lastCone) {
-          const box = new THREE.Box3().setFromObject(lastCone);
-          cone.position.y = (box.max.y + box.min.y) / 2;
-        } else {
-          cone.position.y += 2;
-        }
-        lastCone = cone;
-        cone.castShadow = true;
-        cone.receiveShadow = true;
-        this.add(cone);
-      }
-    }
   }
 
   // snow flowers
